common: Change `MegaId` to `u16`
Since we do collision detection and these ids aren't shared across
machines it doesn't need to be collision-resistant.

+++ common/src/api/mod.rs
@@ -21,7 +21,7 @@ pub mod user;
 pub mod version;
 
 /// A randomly generated id for each mega node.
-pub type MegaId = u64;
+pub type MegaId = u16;
 
 #[derive(Copy, Clone, Debug, PartialEq, Serialize, Deserialize)]
 pub struct MegaIdStruct {
